requirements.txt
PyQt5==5.15.10
face-recognition==1.3.0
opencv-python==4.8.1.78
mysql-connector-python==8.2.0
numpy==1.24.3
dlib==19.24.2
cmake==3.27.5

